Conflicts:
	systems/plants/RigidBodyManipulator.h

git-svn-id: https://svn.csail.mit.edu/locomotion/robotlib/branches/hongkaidev@7926 c9849af7-e679-4ec6-a44e-fc146a885bd3

+++ systems/plants/HandCmex.cpp
@@ -30,7 +30,7 @@ void mexFunction( int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[] ) {
   
   double *q,*qd;
   Map<MatrixXd> *f_ext=NULL;
-  if (mxGetNumberOfElements(prhs[1])!=model->num_dof || mxGetNumberOfElements(prhs[2])!=model->num_dof)
+  if (static_cast<int>(mxGetNumberOfElements(prhs[1]))!=model->num_dof || static_cast<int>(mxGetNumberOfElements(prhs[2]))!=model->num_dof)
     mexErrMsgIdAndTxt("Drake:HandCmex:BadInputs","q and qd must be size %d x 1",model->num_dof);
   q = mxGetPr(prhs[1]);
   qd = mxGetPr(prhs[2]);
